requirements.txt
fpdf2==2.7.4
numpy==1.24.2